Move network benchmarks to benches directory
- Also document new code.
