+++ exonum/Cargo.toml
@@ -64,4 +64,5 @@ default = ["leveldb"]
 leveldb = ["exonum_leveldb"]
 rocksdb = ["exonum_rocksdb"]
 network_benchmarks = []
+long_benchmarks = []
 flame_profile = ["exonum_profiler/nomock", "exonum_flamer"]
