+++ exonum/src/events/mod.rs
@@ -12,9 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#[cfg(test)]
-mod tests;
-
+#[cfg(any(test, feature = "network_benchmarks"))]
+pub mod tests;
 pub mod codec;
 pub mod error;
 pub mod network;
